Authorize task and user_task

+++ app/controllers/tasks_controller.rb
@@ -1,39 +1,53 @@
 class TasksController < ApplicationController
-  load_and_authorize_resource
-  before_action :load_user_subject, only: [:create, :edit, :update, :destroy]
-  before_action :add_task_info, only: [:create]
+  before_action :load_task, only: [:update, :destroy]
+  before_action :load_user_subject_in_course, only: [:edit, :update, :destroy]
   before_action :load_user_course, only: [:edit, :destroy]
+  after_action :verify_authorized
 
   def create
-    if @task.save
-      user_task = @task.user_tasks.find_by user: current_user
-      user_task_service = MailerServices::UserTaskService.new user_task: user_task,
-        status: Settings.status.init
-      @user_task_history = user_task_service.perform
-      flash[:success] = flash_message "created"
+    @task = current_user.tasks.new task_params
+    if authorize @task
+      load_user_subject_in_course
+      if @task.save
+        user_task = @task.user_tasks.find_by user: current_user
+        user_task_service = MailerServices::UserTaskService.new user_task: user_task,
+          status: Settings.status.init
+        @user_task_history = user_task_service.perform
+        flash[:success] = flash_message "created"
+      else
+        flash[:failed] = flash_message "not_created"
+      end
+      redirect_to :back
     else
-      flash[:failed] = flash_message "not_created"
+      redirect_to root_path
     end
-    redirect_to :back
   end
 
   def update
-    @old_status = user_task.status
-    if @task.update_attributes task_params
-      flash[:success] = flash_message "updated"
+    if authorize @task
+      @old_status = user_task.status
+      if @task.update_attributes task_params
+        flash[:success] = flash_message "updated"
+      else
+        flash[:failed] = flash_message "not_updated"
+      end
+      redirect_to :back
     else
-      flash[:failed] = flash_message "not_updated"
+      redirect_to root_path
     end
-    redirect_to :back
   end
 
   def destroy
-    if @task.destroy
-      flash[:success] = flash_message "deleted"
+    if authorize user_task
+      if @task.destroy
+        flash[:success] = flash_message "deleted"
+      else
+        flash[:failed] = flash_message "not_deteletd"
+      end
+      redirect_to :back
     else
-      flash[:failed] = flash_message "not_deteletd"
+      redirect_to root_path
     end
-    redirect_to :back
   end
 
   private
@@ -41,11 +55,7 @@ class TasksController < ApplicationController
     params.require(:task).permit Task::ATTRIBUTES_PARAMS
   end
 
-  def add_task_info
-    @task.create_by_trainee = current_user.is_trainee?
-  end
-
-  def load_user_subject
+  def load_user_subject_in_course
     @user_subject = UserSubject.find_by user: current_user,
       course_subject: @task.course_subject
   end
